fix-10: Managers+Player — 21 задач (MGR-C01/02/03,H01/02/03/04/05,PLR-H01/02/03,M01-07,L01/02/03) isInitializing, SceneLoader fixes, ServiceLocator, ICombatant, SleepSystem fixes, PlayerVisual cleanup (2026-04-11)

+++ UnityProject/Assets/Scripts/Save/SaveManager.cs
@@ -787,7 +787,7 @@ namespace CultivationGame.Save
         public string PlayerId;
         public string Name;
         public int CultivationLevel;
-        public int CurrentQi;
+        public long CurrentQi; // FIX PLR-H02: Changed from int to long — Qi can exceed int.MaxValue (2026-04-11)
         public string CurrentLocationId;
     }
 }
